+++ StuHub/app/controllers/messages_controller.rb
@@ -10,18 +10,11 @@ class MessagesController < ApplicationController
   def create
     @message = Message.new(message_params)
     if @message.save
-      @channel = "/messages/new"
-      @type = params[:message][:channel_id]
-      if params[:message][:channel_type].to_i == 1 # Course Chat
-        @channel = "/courses/#{params[:message][:channel_id]}/messages"
-      elsif params[:message][:channel_type].to_i == 2 # Group Chat
-        @channel = "/groups/#{params[:message][:channel_id]}/messages"
-      elsif params[:message][:channel_type].to_i == 3
-        @channel = "/posts/new"
-      elsif params[:message][:channel_type].to_i == 4 # Course Post
-        @channel = "/courses/#{params[:message][:channel_id]}/posts"
-      elsif params[:message][:channel_type].to_i == 5 # Group Post
-        @channel = "/groups/#{params[:message][:channel_id]}/posts"
+      @channel = "/messages/new" # Type 0
+      if @message.channel_type == 1 # Course Chat
+        @channel = "/courses/#{@message.channel_id}/messages"
+      elsif @message.channel_type == 2 # Group Chat
+        @channel = "/groups/#{@message.channel_id}/messages"
       end
       PrivatePub.publish_to(@channel, message: @message)
     end