+++ StuHub/app/helpers/email_changes_controller_helper.rb
@@ -1,0 +1,2 @@
+module EmailChangesControllerHelper
+end