+++ StuHub/app/models/section.rb
@@ -5,7 +5,6 @@ class Section < ActiveRecord::Base
   has_many :exams, dependent: :destroy
   has_many :instructors, dependent: :destroy
 
-  validates :unique_number, length: {minimum: 1}
   validates :key, length: {minimum: 1}
   validates :code, length: {minimum: 1}
   validates :associated_class_id, presence: true