+++ StuHub/db/migrate/20151120103907_add_current_term_to_institution.rb
@@ -1,0 +1,5 @@
+class AddCurrentTermToInstitution < ActiveRecord::Migration
+  def change
+    add_reference :institutions, :current_term, references: :terms
+  end
+end