+++ StuHub/db/migrate/20151125111027_add_term_order_to_term.rb
@@ -1,0 +1,5 @@
+class AddTermOrderToTerm < ActiveRecord::Migration
+  def change
+    add_column :terms, :term_order, :integer, default: 1
+  end
+end