+++ StuHub/db/migrate/20151126114032_add_database_last_scanned_line_to_term.rb
@@ -1,0 +1,5 @@
+class AddDatabaseLastScannedLineToTerm < ActiveRecord::Migration
+  def change
+    add_column :terms, :database_last_line, :integer, default: 0
+  end
+end