+++ StuHub/test/mailers/previews/course_mailer_preview.rb
@@ -1,0 +1,16 @@
+# Preview all emails at http://localhost:3000/rails/mailers/course_mailer
+class CourseMailerPreview < ActionMailer::Preview
+
+  def joined_course
+    user = User.first
+    course = Course.first
+    CourseMailer.joined_course(user, course)
+  end
+
+  def dropped_course
+    user = User.first
+    course = Course.first
+    CourseMailer.dropped_course(user, course)
+  end
+
+end